+++ src/Symfony/Component/Console/Helper/HelperSet.php
@@ -69,16 +69,25 @@ class HelperSet implements \IteratorAggregate
         return $this->helpers[$name];
     }
 
+    /**
+     * @deprecated since Symfony 5.4
+     */
     public function setCommand(Command $command = null)
     {
+        trigger_deprecation('symfony/console', '5.4', 'Method "%s()" is deprecated.', __METHOD__);
+
         $this->command = $command;
     }
 
     /**
      * Gets the command associated with this helper set.
+     *
+     * @deprecated since Symfony 5.4
      */
     public function getCommand(): Command
     {
+        trigger_deprecation('symfony/console', '5.4', 'Method "%s()" is deprecated.', __METHOD__);
+
         return $this->command;
     }